Fix for issue #8
Add new reindexing target "UNAVAILABLE" which selects all indexes that
are not ENABLED or DISABLED. For the graph indexes the state of each
index property key is taken into account, e.g. if the index status for
at least one property is INSTALLED or REGISTERED but there is no
property with the state DISABLED, then the index will be selected for
update (recovery / reindexing). This option can be used when a previous
graph schema update fails for any reason. It only selects the indexes
defined in the schema.

Signed-off-by: Nikolai Grigoriev <ngrigoriev@newforma.com>